django_dashboard/requirements.txt
Django>=4.2
google-cloud-bigquery>=3.0.0
pandas>=2.0.0
plotly>=5.18.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
mlflow>=2.9.0
